+++ Entities/Master/POSCustomer.cs
@@ -6,7 +6,7 @@ namespace POS_API.Entities.Master
     {
         [Key]
         [Required]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
 
         public bool IsActive { get; set; } = true;
@@ -19,17 +19,17 @@ namespace POS_API.Entities.Master
         public DateTime CreatedAt { get; set; }
 
 
-        public long? CreatedBy { get; set; }
+        public int? CreatedBy { get; set; }
 
 
         public DateTime? UpdatedAt { get; set; }
 
 
-        public long? UpdatedBy { get; set; }
+        public int? UpdatedBy { get; set; }
         public DateTime? DeletedAt { get; set; }
 
 
-        public long? DeletedBy { get; set; }
+        public int? DeletedBy { get; set; }
         [Required]
         [MaxLength(50)]
         public string? CustomerCode { get; set; }
@@ -49,6 +49,6 @@ namespace POS_API.Entities.Master
 
         [MaxLength(300)]
         public string? Address { get; set; }
-        public long? CompanyId { get; set; }
+        public int? CompanyId { get; set; }
     }
 }
