+++ Entities/Master/POSProductBatch.cs
@@ -9,8 +9,8 @@ namespace POS_API.Entities.Master
     {
         [Key]
         [Required]
-        public long Id { get; set; }
-        public long? CompanyId { get; set; }
+        public int Id { get; set; }
+        public int? CompanyId { get; set; }
 
         public bool IsActive { get; set; } = true;
 
@@ -19,16 +19,16 @@ namespace POS_API.Entities.Master
         [Required]
         public DateTime CreatedAt { get; set; }
 
-        public long? CreatedBy { get; set; }
+        public int? CreatedBy { get; set; }
 
         public DateTime? UpdatedAt { get; set; }
 
-        public long? UpdatedBy { get; set; }
+        public int? UpdatedBy { get; set; }
         public DateTime? DeletedAt { get; set; }
 
-        public long? DeletedBy { get; set; }
+        public int? DeletedBy { get; set; }
         [Required]
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
         [ForeignKey(nameof(ProductId))]
         public POSProduct? Product { get; set; }
